+++ test-utils/test-utils-core/src/test/java/uk/gov/justice/services/test/utils/core/http/HttpResponsePollerTest.java
@@ -1,45 +1,52 @@
 package uk.gov.justice.services.test.utils.core.http;
 
+import static java.util.Collections.singletonMap;
 import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.OK;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.core.IsNull.nullValue;
+import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
 import uk.gov.justice.services.test.utils.core.rest.RestClient;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import javax.ws.rs.core.Response;
 
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
 public class HttpResponsePollerTest {
 
-    private final static String EXPECTED_MESSAGE_404 = "Failed to get 404 response from 'http://localhost:8080/poll/condition' after 5 attempts. Status code: 404";
-    private final static String EXPECTED_MESSAGE_200 = "Failed to get 200 response from 'http://localhost:8080/poll/condition' after 5 attempts. Status code: 200";
+    private final static String EXPECTED_MESSAGE_404 = "Failed to match response conditions from http://localhost:8080/poll/condition, after 5 attempts, with status code: 404";
+    private final static String EXPECTED_MESSAGE_200 = "Failed to match response conditions from http://localhost:8080/poll/condition, after 5 attempts, with status code: 200";
+    private final static String EXPECTED_MESSAGE_RESULT = "Failed to match result conditions from http://localhost:8080/poll/condition, after 5 attempts, with result: ";
+
     private final String URL = "http://localhost:8080/poll/condition";
     private final String MEDIA_TYPE = "application/vnd.poll.condition+json";
-    private final int DELAY_IN_MILLIS = 100;
+
     @Rule
     public ExpectedException expectedException = ExpectedException.none();
+
     @Mock
     private RestClient restClient;
+
     @Mock
     private Response response;
 
+    @InjectMocks
     private HttpResponsePoller poller;
 
-    @Before
-    public void setup() throws Exception {
-        poller = new HttpResponsePoller(restClient);
-    }
-
     @Test
-    public void shouldPollOnFoundResponseUntilSuccessful() {
+    public void shouldReturnResultEntityIfFound() {
 
         final String responseText = "RESPONSE";
 
@@ -47,7 +54,9 @@ public class HttpResponsePollerTest {
         when(response.getStatus()).thenReturn(OK.getStatusCode());
         when(response.readEntity(String.class)).thenReturn(responseText);
 
-        poller.pollUntilFound(URL, MEDIA_TYPE);
+        final String result = poller.pollUntilFound(URL, MEDIA_TYPE);
+
+        assertThat(result, is(responseText));
     }
 
     @Test
@@ -62,23 +71,25 @@ public class HttpResponsePollerTest {
         expectedException.expect(AssertionError.class);
         expectedException.expectMessage(EXPECTED_MESSAGE_404);
 
-        poller.pollUntilFound(URL, MEDIA_TYPE);
+        final String result = poller.pollUntilFound(URL, MEDIA_TYPE);
+
+        assertThat(result, is(responseText));
     }
 
     @Test
-    public void shouldPollOnNotFoundResponseUntilSuccessful() {
-
-        final String responseText = "RESPONSE";
+    public void shouldReturnNullValueIfNotFound() {
 
         when(restClient.query(URL, MEDIA_TYPE)).thenReturn(response);
         when(response.getStatus()).thenReturn(NOT_FOUND.getStatusCode());
-        when(response.readEntity(String.class)).thenReturn(responseText);
+        when(response.readEntity(String.class)).thenReturn(null);
 
-        poller.pollUntilNotFound(URL, MEDIA_TYPE);
+        final String result = poller.pollUntilNotFound(URL, MEDIA_TYPE);
+
+        assertThat(result, nullValue());
     }
 
     @Test
-    public void shouldFailPollOnNotFoundWithFoundResponse() {
+    public void shouldPollAndFailIfFound() {
 
         final String responseText = "RESPONSE";
 
@@ -101,11 +112,13 @@ public class HttpResponsePollerTest {
         when(response.getStatus()).thenReturn(OK.getStatusCode());
         when(response.readEntity(String.class)).thenReturn(responseText);
 
-        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, DELAY_IN_MILLIS, response -> response.equals(responseText));
+        final String result = poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, responseText::equals);
+
+        assertThat(result, is(responseText));
     }
 
     @Test
-    public void shouldPollAndFailOnNullResponseOnCondition() {
+    public void shouldPollAndFailIfNullResponseEntity() {
 
         final String responseConditionMet = "Condition Met";
 
@@ -114,11 +127,10 @@ public class HttpResponsePollerTest {
         when(response.readEntity(String.class)).thenReturn(null);
 
         expectedException.expect(AssertionError.class);
-        expectedException.expectMessage(EXPECTED_MESSAGE_200);
+        expectedException.expectMessage(EXPECTED_MESSAGE_RESULT);
 
-        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, DELAY_IN_MILLIS, response -> response.equals(responseConditionMet));
+        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, responseConditionMet::equals);
     }
-
 
     @Test
     public void shouldPollAndFailWhenConditionNotMetAfter5Retries() {
@@ -133,7 +145,7 @@ public class HttpResponsePollerTest {
         expectedException.expect(AssertionError.class);
         expectedException.expectMessage(EXPECTED_MESSAGE_404);
 
-        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, DELAY_IN_MILLIS, response -> response.equals(responseConditionMet));
+        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, responseConditionMet::equals);
     }
 
     @Test
@@ -150,7 +162,69 @@ public class HttpResponsePollerTest {
                 .thenReturn(responseConditionNotMet)
                 .thenReturn(responseConditionMet);
 
-        poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, DELAY_IN_MILLIS, response -> response.equals(responseConditionMet));
+        final String result = poller.pollUntilFoundWithCondition(URL, MEDIA_TYPE, responseConditionMet::equals);
+
+        assertThat(result, is(responseConditionMet));
+    }
+
+    @Test
+    public void shouldPollUntilJsonObjectWithMatchingValuesIsFound() throws Exception {
+
+        final String matchingJsonObject = "{\"name\": \"test\",\"email\": \"test@email.co.uk\"}";
+        final String unMatchingJsonObject = "{\"name\": \"no match\"}";
+
+        final Map<String, String> matchValues = new HashMap<>();
+        matchValues.put("name", "test");
+        matchValues.put("email", "test@email.co.uk");
+
+        when(restClient.query(URL, MEDIA_TYPE)).thenReturn(response);
+        when(response.getStatus()).thenReturn(OK.getStatusCode());
+
+        when(response.readEntity(String.class))
+                .thenReturn(unMatchingJsonObject)
+                .thenReturn(matchingJsonObject);
+
+        final String result = poller.pollUntilJsonObjectFoundWithValues(URL, MEDIA_TYPE, matchValues);
+
+        assertThat(result, is(matchingJsonObject));
+    }
+
+    @Test
+    public void shouldFailIfValueIsNotFound() throws Exception {
+
+        final String unMatchingJsonObject = "{\"name\": \"no match\"}";
+        final Map<String, String> matchValues = singletonMap("name", "test");
+
+        when(restClient.query(URL, MEDIA_TYPE)).thenReturn(response);
+        when(response.getStatus()).thenReturn(OK.getStatusCode());
+
+        when(response.readEntity(String.class))
+                .thenReturn(unMatchingJsonObject);
+
+        expectedException.expect(AssertionError.class);
+        expectedException.expectMessage(EXPECTED_MESSAGE_RESULT);
+
+        poller.pollUntilJsonObjectFoundWithValues(URL, MEDIA_TYPE, matchValues);
+    }
+
+    @Test
+    public void shouldFailIfASingleValueOfMultipleValuesIsNotFound() throws Exception {
+
+        final String matchingJsonObject = "{\"name\": \"test\",\"email\": \"test@email.co.uk\"}";
+        final Map<String, String> matchValues = new HashMap<>();
+        matchValues.put("name", "test");
+        matchValues.put("email", "notmatch@email.co.uk");
+
+        when(restClient.query(URL, MEDIA_TYPE)).thenReturn(response);
+        when(response.getStatus()).thenReturn(OK.getStatusCode());
+
+        when(response.readEntity(String.class))
+                .thenReturn(matchingJsonObject);
+
+        expectedException.expect(AssertionError.class);
+        expectedException.expectMessage(EXPECTED_MESSAGE_RESULT);
+
+        poller.pollUntilJsonObjectFoundWithValues(URL, MEDIA_TYPE, matchValues);
     }
 
 }